Add retries for wifi setup
Instead of locking the mcu if we fail to initalize the CYW43 or
conecting to a wifi network, we should retry, a definable number of
times, and return to the application code if possible.

Testing: Forced a retry of wifi connect by using a non existent network,
with 3 retries and delay of 1000ms between retries.